+++ lib/src/models/models.dart
@@ -1,5 +1,6 @@
 export './Battery.dart';
 export './Climate.dart';
+export './Command.dart';
 export './Software.dart';
 export './Statistics.dart';
 export './Vehicle.dart';