+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 - Added models for vehicles and software
 - Added small example
 - Added test for some of the vehicle values and some of the software values
+
+## 0.0.2
+
+- Added model for vehicle commands
+- Added the update command for a vehicle in VehicleAPI
+- Added an enum for the state a car can be in
+- Added the command for getting vehicle state as an enum